+++ bin-proto/src/lib.rs
@@ -387,6 +387,19 @@ pub use bitstream_io::{BigEndian, BitRead, BitWrite, Endianness, LittleEndian};
 /// #[derive(BitDecode, BitEncode)]
 /// pub struct Struct(#[codec(pad_after = 3)] u8);
 /// ```
+///
+/// ## `#[codec(magic = <expr>)]`
+/// - Applies to: fields, structs
+/// - `<expr>`: Must evaluate to `&[u8; _]`
+///
+/// Indicates that the value must be present immediately preceding the field or struct.
+///
+/// ```
+/// # use bin_proto::{BitDecode, BitEncode};
+/// #[derive(BitDecode, BitEncode)]
+/// #[codec(magic = &[0x01, 0x02, 0x03])]
+/// pub struct Magic(#[codec(magic = b"123")] u8);
+/// ```
 #[cfg_attr(docsrs, doc(cfg(feature = "derive")))]
 #[cfg(feature = "derive")]
 pub use bin_proto_derive::{BitDecode, BitEncode};